Clean up and moving of classes to correct domain
